+++ packages/blade/src/components/SpotlightPopoverTour/TourDocs.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
+import { BasicExample } from './docs/examples';
 import StoryPageWrapper from '~utils/storybook/StoryPageWrapper';
 import { Code, Heading, Text, Title } from '~components/Typography';
-import { Sandbox, SandboxHighlighter } from '~utils/storybook/Sandbox/SandpackEditor';
 import { Box } from '~components/Box';
 import type { TableData } from '~components/Table';
 import {
@@ -16,6 +16,8 @@ import {
 import { ScrollLink } from '~utils/storybook/ScrollLink';
 import { InfoIcon } from '~components/Icons';
 import { Popover, PopoverInteractiveWrapper } from '~components/Popover';
+import { SandboxHighlighter } from '~utils/storybook/Sandbox/SandpackEditor';
+import { Sandbox } from '~utils/storybook/Sandbox/StackblitzEditor/Sandbox';
 
 type Item = {
   id: string;
@@ -237,70 +239,7 @@ const TourDocs = (): React.ReactElement => {
       }}
     >
       <Title>Usage</Title>
-      <Sandbox>
-        {`
-        import React from 'react';
-        import { 
-          SpotlightPopoverTour,
-          SpotlightPopoverTourStep,
-          SpotlightPopoverTourFooter,
-          Box,
-          Text,
-          Button
-        } from '@razorpay/blade/components';
-        
-        function App(): React.ReactElement {
-          const [activeStep, setActiveStep] = React.useState(0);
-          const [isOpen, setIsOpen] = React.useState(false);
-        
-          return (
-            <Box>
-              <Button
-                marginBottom="spacing.9"
-                onClick={() => {
-                  setIsOpen((prev) => !prev);
-                }}
-              >
-                {isOpen ? 'Tour In Progress' : 'Start Tour'}
-              </Button>
-              <SpotlightPopoverTour
-                steps={steps}
-                isOpen={isOpen}
-                activeStep={activeStep}
-                onFinish={() => {
-                  console.log('finished');
-                  setActiveStep(0);
-                  setIsOpen(false);
-                }}
-                onOpenChange={({ isOpen }) => {
-                  console.log('open change', isOpen);
-                  setIsOpen(isOpen);
-                }}
-                onStepChange={(step) => {
-                  console.log('step change', step);
-                  setActiveStep(step);
-                }}
-              >
-                <Box width="100%" display="flex" gap="spacing.4">
-                  <SpotlightPopoverTourStep name="step-1">
-                    <Box padding="spacing.4" backgroundColor="brand.gray.400.lowContrast">
-                      <Text>Step 1 </Text>
-                    </Box>
-                  </SpotlightPopoverTourStep>
-                  <SpotlightPopoverTourStep name="step-2">
-                    <Box padding="spacing.4" backgroundColor="brand.gray.400.lowContrast">
-                      <Text>Step 2 </Text>
-                    </Box>
-                  </SpotlightPopoverTourStep>
-                </Box>
-              </SpotlightPopoverTour>
-            </Box>
-          );
-        }
-
-        export default App;
-      `}
-      </Sandbox>
+      <Sandbox>{BasicExample}</Sandbox>
       <Title>iOS Safari Specific Setup</Title>
       <Text marginTop="spacing.5">
         When using BottomSheet or SpotlightPopoverTour, Make sure to set a width/height to the
