+++ packages/blade/src/components/SpotlightPopoverTour/docs/Tour.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/addon-docs';
-import { TourDocs } from './TourDocs';
+import { TourDocs } from '../TourDocs';
 
 <Meta title="Components/SpotlightPopoverTour/API" />
