+++ packages/blade/src/components/SpotlightPopoverTour/docs/Tour.stories.tsx
@@ -3,14 +3,15 @@ import type { Meta, ComponentStory } from '@storybook/react';
 import React from 'react';
 import { Title } from '@storybook/addon-docs';
 import isChromatic from 'chromatic';
+import { SpotlightPopoverTourStep } from '../TourStep';
+import { SpotlightPopoverTourFooter } from '../TourFooter';
+import { SpotlightPopoverTour } from '../..';
 import type {
   SpotlightPopoverStepRenderProps,
   SpotlightPopoverTourProps,
   SpotlightPopoverTourSteps,
-} from './types';
-import { SpotlightPopoverTourStep } from './TourStep';
-import { SpotlightPopoverTourFooter } from './TourFooter';
-import { SpotlightPopoverTour } from '.';
+} from '../types';
+import { BasicExample } from './examples';
 import { Button } from '~components/Button';
 import { Box } from '~components/Box';
 import { Code, Text } from '~components/Typography';
@@ -36,137 +37,7 @@ const Page = (): React.ReactElement => {
       }}
     >
       <Title>Usage</Title>
-      <Sandbox>
-        {`
-        import React from 'react';
-        import { 
-          SpotlightPopoverTour,
-          SpotlightPopoverTourStep,
-          SpotlightPopoverTourFooter,
-          Box,
-          Text,
-          Button
-        } from '@razorpay/blade/components';
-        import type {
-          SpotlightPopoverStepRenderProps,
-          SpotlightPopoverTourSteps,
-        } from '@razorpay/blade/components';
-
-        const CustomTourFooter = ({
-          activeStep,
-          totalSteps,
-          goToNext,
-          goToPrevious,
-          stopTour,
-        }: SpotlightPopoverStepRenderProps) => {
-          const isLast = activeStep === totalSteps - 1;
-          const isFirst = activeStep === 0;
-          return (
-            <SpotlightPopoverTourFooter
-              activeStep={activeStep}
-              totalSteps={totalSteps}
-              actions={{
-                primary: isLast
-                  ? {
-                      text: 'Done',
-                      onClick: stopTour,
-                    }
-                  : {
-                      text: 'Next',
-                      onClick: goToNext,
-                    },
-                secondary: isFirst
-                  ? undefined
-                  : {
-                      text: 'Prev',
-                      onClick: goToPrevious,
-                    },
-              }}
-            />
-          );
-        };
-
-        const steps: SpotlightPopoverTourSteps = [
-          {
-            name: 'step-1',
-            title: 'Step 1',
-            content: () => {
-              return (
-                <Box>
-                  <Text>This is step 1</Text>
-                </Box>
-              );
-            },
-            placement: 'top',
-            footer: CustomTourFooter,
-          },
-          {
-            name: 'step-2',
-            title: 'Step 2',
-            content: () => {
-              return (
-                <Box>
-                  <Text>This is step 2</Text>
-                </Box>
-              );
-            },
-            placement: 'bottom',
-            footer: CustomTourFooter,
-          },
-        ];
-
-        function App(): React.ReactElement {
-          const [activeStep, setActiveStep] = React.useState(0);
-          const [isOpen, setIsOpen] = React.useState(false);
-        
-          return (
-            <Box>
-              <Button
-                marginBottom="spacing.9"
-                onClick={() => {
-                  setIsOpen((prev) => !prev);
-                }}
-              >
-                {isOpen ? 'Tour In Progress' : 'Start Tour'}
-              </Button>
-              <SpotlightPopoverTour
-                steps={steps}
-                isOpen={isOpen}
-                activeStep={activeStep}
-                onFinish={() => {
-                  console.log('finished');
-                  setActiveStep(0);
-                  setIsOpen(false);
-                }}
-                onOpenChange={({ isOpen }) => {
-                  console.log('open change', isOpen);
-                  setIsOpen(isOpen);
-                }}
-                onStepChange={(step) => {
-                  console.log('step change', step);
-                  setActiveStep(step);
-                }}
-              >
-                <Box width="100%" display="flex" gap="spacing.4">
-                  <SpotlightPopoverTourStep name="step-1">
-                    <Box padding="spacing.4" backgroundColor="brand.gray.400.lowContrast">
-                      <Text>Step 1 </Text>
-                    </Box>
-                  </SpotlightPopoverTourStep>
-                  <SpotlightPopoverTourStep name="step-2">
-                    <Box padding="spacing.4" backgroundColor="brand.gray.400.lowContrast">
-                      <Text>Step 2 </Text>
-                    </Box>
-                  </SpotlightPopoverTourStep>
-                </Box>
-              </SpotlightPopoverTour>
-            </Box>
-          );
-        }
-
-        export default App;
-      `}
-      </Sandbox>
+      <Sandbox>{BasicExample}</Sandbox>
       <Title>iOS Safari Specific Setup</Title>
       <Text marginTop="spacing.5">
         When using BottomSheet or SpotlightPopoverTour, Make sure to set a width/height to the
